x86: Preliminary LAPIC timer support
Added support for I/O APIC (currently just mapping PIC IRQs to what we
are already mapping the PIC to), and finished minimal support for the
APIC.

The LAPIC timer support does not currently calibrate against the PIT, it
just assumes a speed of 10 MHz (randomly chosen for testing). So it may
run too fast or too slow on different systems (and on no system will it
run at a reasonable rate).